pyproject.toml
[tool.poetry]
name = "currency-rate"
version = "0.1.0"
description = ""
authors = ["krasniy-1889 <neytirri@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django = "^5.0.2"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
argon2-cffi = "^23.1.0"
celery = "^5.3.6"
httpx = "^0.26.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
